feat(, ): Add webpack and create home Django app
Modified webpack to optimize its use with django-webpack-loader. Some of the old contents that were

removed are now in a comment and can be added back to the extent that they don't damage the

environment. Also created a new "home" app to be the location that django serves up the frontend.