feat(, ): Serve webpack bundle properly to Django using 'django-webpack-loader,' rewr
Using the 'django-webpack-loader' on the Python end with a retooled webpack.config.js file, the

Django '/home' route now serves the React/Redux app. A few things to keep in mind. '/' route will

now redirect to '/home.' Accordingly the endpoints in react-router now begin with '/home' (e.g.

'home/register'). An issue that needs to be resolved is that the user cannot directly navigate to a

specific route in the React as Django will try to find the appropriate view and throw a 404.

Additionally, the webpack.config.js file should be reviewed by other members on the team before

being merged into the project.